refactor(temporal): remove docs-groom workflow
The docs-groom workflow accumulated too many class-of-bugs (JSON parse,
schema drift, prompt drift, auth, branch hygiene, PR layout) and the
queue of draft PRs requires human review anyway, so the maintenance
load isn't justified. Removing the feature in one PR.

Deletes the workflow, activities, shared modules, dev tooling, the
schedule registration, the Prometheus metrics, the PrometheusRule alert
group, and the Grafana dashboard panels. Extracts the one helper
pr-agent still uses (parseClaudeResultMessage) into shared/claude-result.ts.
Strips docs-groom-specific framing from worker.ts comments and CLAUDE.md.

Keeps orthogonal fixes that landed during the docs-groom debugging arc:
the workflow-bundle smoke test, the CI image-scope filter, the
version-commit-back filter, the ANTHROPIC_API_KEY removal, the
GIT_AUTHOR_* env vars (still useful for any future git-committing
activity), and the temporal-worker-zfs-exec daemonsets:get permission.

After deploy, the docs-groom-daily schedule will be deregistered on
worker startup. Open docs-groom PRs and labels will be cleaned up
manually post-merge.
