+++ packages/temporal/src/schedules/register-schedules.ts
@@ -95,23 +95,13 @@ const SCHEDULES: ScheduleDefinition[] = [
     id: "velero-orphan-audit",
     workflowType: "runVeleroOrphanAuditWorkflow",
     args: [],
-    // 03:30 PT — staggered after zfs-maintenance and before docs-groom so the
-    // audit captures a stable post-backup state.
+    // 03:30 PT — staggered after zfs-maintenance so the audit captures a
+    // stable post-backup state.
     cronExpression: "30 3 * * *",
     taskQueue: TASK_QUEUES.DEFAULT,
     overlap: ScheduleOverlapPolicy.SKIP,
     workflowExecutionTimeout: "15 minutes",
     memo: "Daily Velero orphan ZFS snapshot detection — emits Prometheus metrics for the orphan-snapshot pathology (see packages/docs/decisions/2026-05-05_velero-orphan-snapshot-prevention.md)",
-  },
-  {
-    id: "docs-groom-daily",
-    workflowType: "runDocsGroomAudit",
-    args: [],
-    cronExpression: "30 6 * * *",
-    taskQueue: TASK_QUEUES.DEFAULT,
-    overlap: ScheduleOverlapPolicy.SKIP,
-    workflowExecutionTimeout: "3 hours",
-    memo: "Daily packages/docs grooming via claude -p; opens draft PRs for easy/medium tasks",
   },
   {
     id: "golink-sync",
