--- packages/temporal/src/shared/docs-groom-prompts.ts
@@ -1,168 +1,0 @@
-/**
- * Prompts for the docs-groom workflow. Kept as exported constants so they
- * are reviewable in one place and can be unit-tested if needed.
- *
- * Both prompts assume Claude is invoked via
- * `claude -p --output-format json --json-schema <SchemaFromZod>` with
- * `--allowed-tools "Read,Write,Edit,Glob,Grep"` and
- * `--permission-mode acceptEdits`. `--json-schema` enforces structured
- * output upstream — the prompt only describes WHAT to produce, never HOW
- * to format it.
- */
-
-export const GROOM_PROMPT = `You are running as part of a daily automated grooming pass over the
-\`packages/docs/\` directory of the shepherdjerred/monorepo. You are working
-in a fresh git worktree on a feature branch off origin/main. You can read,
-write, and edit files freely within the worktree.
-
-# Authoritative conventions
-
-Before doing anything else, read these two files and follow them strictly:
-- \`packages/docs/CLAUDE.md\` — doc-system conventions
-- \`packages/docs/index.md\` — the table of contents you must keep in sync
-
-Key rules from CLAUDE.md you MUST respect:
-- Naming: \`<yyyy-mm-dd>_kebab-case.md\`
-- Categories: \`architecture/\`, \`patterns/\`, \`decisions/\`, \`guides/\`, \`plans/\`
-- Plans must have a \`## Status\` section near the top
-- Archive don't delete: outdated docs go to \`archive/<technology>/\` or \`archive/superseded/\`
-- No root-level docs (every doc lives in a subdirectory)
-- Update \`index.md\` after any moves, additions, or removals
-
-# Your two responsibilities
-
-## 1. DO the easy in-place grooming yourself
-
-Edit files directly. Examples of what to do inline:
-- Move stale docs into \`archive/\` (with the appropriate subdirectory)
-- Move superseded plans into \`archive/superseded/\`
-- Add missing \`## Status\` sections to plans
-- Update \`index.md\` to reflect any moves you make
-- Fix broken intra-doc links (relative markdown links that 404)
-- Fix naming-convention violations (rename files to \`<date>_kebab-case.md\`)
-- Move any root-level docs into subdirectories
-- Anything else that is mechanical and low-risk
-
-Be conservative — when in doubt, don't groom inline; surface as a task instead.
-
-## 2. IDENTIFY larger improvements you did NOT do
-
-Bigger changes that need their own focused PR. Examples:
-- Rewriting an outdated doc against current code/state
-- Splitting an oversized doc into smaller focused ones
-- Drafting a missing pattern doc that the codebase clearly needs
-- Verifying a plan marked "Implemented" against actual code (search the
-  repo for the package/script/config it describes; if missing, the status
-  is wrong)
-- Expanding a sparse decisions/ doc with proper context
-
-Cap: at most 10 tasks total in your output list. Prioritize the most
-valuable. Skip anything you already completed inline.
-
-Difficulty rubric:
-- \`easy\`: < 30 min wall clock for a focused Claude session, < 100 LOC
-  changed, single-file or near-mechanical
-- \`medium\`: 30 min – 2 h, multi-file but bounded scope
-- \`hard\`: requires deep code investigation, cross-package context, or a
-  large rewrite. These will NOT be auto-implemented.
-
-# Output
-
-Return:
-- \`summary\`: one paragraph describing what you groomed inline (becomes the grooming PR body)
-- \`groomedFiles\`: paths you edited inline
-- \`tasks\`: larger improvements you did NOT do, each with \`title\`, \`slug\` (kebab-case), \`description\` (≥20 chars, with enough context for a follow-up Claude session to implement WITHOUT rerunning the audit), \`difficulty\`, \`files\`, \`category\`
-
-\`category\` MUST be exactly one of these strings (lowercase, kebab-case):
-- \`stale\` — the doc is outdated and should be moved to \`archive/\` or rewritten
-- \`broken-link\` — relative markdown links inside the doc 404
-- \`status-rot\` — a plan's \`## Status\` section is missing or contradicts current code
-- \`index-drift\` — \`packages/docs/index.md\` doesn't reflect the actual contents of \`packages/docs/\`
-- \`unverified-implemented\` — a plan claims "Implemented" but the referenced code/config doesn't exist
-- \`rewrite\` — the doc needs a substantive rewrite against current state
-- \`split\` — the doc has grown too large and should be broken into focused docs
-- \`other\` — none of the above; use sparingly
-
-Do NOT invent new category values (e.g. "architecture", "guide", "doc",
-"plan"). Those are *folder* names in \`packages/docs/\`, not task categories.
-The output schema is enforced by \`claude --json-schema\`; values outside
-the list above will be rejected.
-
-If you made no inline edits and identified no tasks, return empty \`groomedFiles\` and \`tasks\` with summary "No grooming or follow-up tasks needed."
-
-# JSON syntax — CRITICAL
-
-The output is parsed as **strict JSON** — one shot, no retry. Any
-syntax error fails the entire grooming run.
-
-- All strings MUST use double-quote delimiters (\`"..."\`).
-- Do NOT use backtick (\`\` \`...\` \`\`) or single-quote (\`'...'\`)
-  delimiters anywhere in the output. Those are JavaScript / Python
-  syntax, not JSON.
-- Inside string values, escape literal double quotes as \`\\"\` and
-  literal backslashes as \`\\\\\`. Backticks inside strings are fine.
-- Title MUST be ≤120 characters. Description MUST be ≤2000 characters.
-  Don't produce values longer than that — the parser will truncate
-  them but you should not rely on that.
-`;
-
-export function buildImplementPrompt(input: {
-  title: string;
-  slug: string;
-  description: string;
-  difficulty: string;
-  category: string;
-  files: string[];
-  branch: string;
-}): string {
-  return `You are running as part of an automated grooming pass over the
-shepherdjerred/monorepo. You are working in a fresh git worktree on
-feature branch \`${input.branch}\` off origin/main. Your job: implement
-exactly the following ONE task, no more, no less.
-
-# Task
-
-- **Title**: ${input.title}
-- **Slug**: ${input.slug}
-- **Difficulty**: ${input.difficulty}
-- **Category**: ${input.category}
-- **Expected files**: ${input.files.length === 0 ? "(unspecified — use your judgement)" : input.files.join(", ")}
-
-# Description
-
-${input.description}
-
-# Conventions
-
-\`packages/docs/CLAUDE.md\` is authoritative for doc structure & naming.
-Read it before editing.
-
-- Naming: \`<yyyy-mm-dd>_kebab-case.md\`
-- Plans need a \`## Status\` section near the top
-- Archive don't delete: outdated docs go to \`archive/\`
-- Update \`packages/docs/index.md\` if you move or rename any doc
-- Edits outside \`packages/docs/\` are allowed when this task explicitly
-  requires them (e.g. updating a stale code reference). Keep the diff
-  scoped to this one task — no incidental refactors. No new files outside
-  \`packages/docs/\` unless the task description names them.
-
-# Output
-
-After completing the task, return:
-- \`summary\`: one paragraph rationale describing what you changed and why (becomes the PR body)
-- \`filesChanged\`: paths you actually changed (must be non-empty)
-
-# JSON syntax — CRITICAL
-
-The output is parsed as **strict JSON** — one shot, no retry. Any
-syntax error fails this task.
-
-- All strings MUST use double-quote delimiters (\`"..."\`).
-- Do NOT use backtick (\`\` \`...\` \`\`) or single-quote (\`'...'\`)
-  delimiters anywhere. No Markdown formatting like \`**summary**: ...\` —
-  emit the raw JSON object only, no prose, no \`---\` headers, no
-  fenced code blocks.
-- Inside string values, escape literal double quotes as \`\\"\` and
-  literal backslashes as \`\\\\\`.
-`;
-}
